refactor: move helpers to documents package
