+++ packages/documents/lib/meta.js
@@ -158,6 +158,37 @@ const getEstimatedConsumptionMinutes = (doc, estimatedReadingMinutes) =>
     ? Math.round(doc.meta.audioSource.durationMs / 1000 / 60)
     : estimatedReadingMinutes
 
+// _meta is present on unpublished docs
+// { repo { publication { commit { document } } } }
+const getRepoIdsForDoc = (doc, includeParents) => ([
+  doc.meta?.repoId || doc._meta?.repoId,
+  includeParents && getRepoId(
+    doc.meta?.format || doc._meta?.format
+  ).repoId
+].filter(Boolean))
+
+const getTemplate = (doc) =>
+  doc.meta?.template || doc._meta?.template
+
+const getAuthorUserIds = (doc, { loaders }, credits) =>
+  Promise.map(
+    (doc?.meta?.credits || doc?._meta?.credits || credits)
+      .filter(c => c.type === 'link'),
+    async ({ url }) => {
+      if (url.startsWith('/~')) {
+        const idOrUsername = url.substring(2)
+        if (isUuid(idOrUsername)) {
+          return idOrUsername
+        } else {
+          return loaders.User.byUsername.load(idOrUsername)
+            .then(u => u?.id)
+        }
+      } else {
+        console.warn(`invalid author link: ${url} doc: ${(doc.meta || doc._meta)?.repoId}`)
+      }
+    }
+  ).then(userIds => userIds.filter(Boolean))
+
 /**
  * Prepares meta information and resolves linked documents in meta which are
  * not available in original {doc.content.meta} fields.
@@ -205,5 +236,8 @@ const getMeta = doc => {
 
 module.exports = {
   getMeta,
-  getWordsPerMinute
+  getWordsPerMinute,
+  getRepoIdsForDoc,
+  getTemplate,
+  getAuthorUserIds
 }
